requirements.txt
uvicorn
fastapi
sqlalchemy
pydantic
typing
pycryptodome
pymysql
cryptography
prometheus_fastapi_instrumentator